chore: convert demo to ts
