+++ webpack.config.js
@@ -11,7 +11,7 @@ const inDev = NODE_ENV === 'development'
 module.exports = {
   mode: inDev ? 'development' : 'production',
   devtool: 'source-map',
-  entry: [path.resolve(__dirname, 'demo/src/index.js')],
+  entry: [path.resolve(__dirname, 'demo/index.tsx')],
   output: {
     path: path.resolve(__dirname, 'build'),
     filename: '[name].[chunkhash].js',
@@ -22,7 +22,7 @@ module.exports = {
     new HTMLWebpackPlugin({
       alwaysWriteToDisk: true,
       minify: true,
-      template: path.resolve(__dirname, 'demo/src/index.html'),
+      template: path.resolve(__dirname, 'demo/index.html'),
       filename: path.resolve(__dirname, 'build/index.html'),
     }),
     new HTMLWebpackHarddiskPlugin(),
@@ -30,9 +30,9 @@ module.exports = {
   module: {
     rules: [
       {
-        test: /\.js$/,
+        test: /\.ts(x)?$/,
         include: /demo/,
-        loader: 'babel-loader',
+        loader: 'ts-loader',
       },
     ],
   },
